+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 The format is based on and uses the types of changes according to [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- Update options version in Issue Templates files: 1_bug_report.yml
+- Change Write-Verbose CmdLet to Write-Output in scripts\Modules\sps.util.psm1
+
+### Fixed
+
+scripts\SPSUpdate.ps1 and scripts\Modules\util.psm1
+
+- Resolve Cannot add Scheduled Task SPSUpdate-FullScript in SharePoint Task Path ([issue #2](https://github.com/luigilink/SPSUpdate/issues/2)).
+
 ## [1.0.0] - 2023-11-20
 
 ### Changed